Move encoding stuff into separate storage package
